Convert to use some of the new jdt.core.manipulation refactorings
- delete local CompilationUnitRewriteOperationsFix and replace with
  jdt.core.manipulation CompilationUnitRewriteOperationsFixCore
- delete local IProposableFix and replace with jdt.core.manipulation version
- delete local CompilationUnitRewrite and replace with jdt.core.manipulation version
- delete local ImportRemover and replace with jdt.core.manipulation version
- update all users of above classes and switch over to use ICleanUpFixCore,
  LinkedPositionGroupCore, and LinkedProposalModelCore from jdt.core.manipulation
